+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.3.0] - 2019-07-18
+- [Changed] Replace UIKitForMac to macCatalyst to reflect the changes in beta 4
+- [Changed] Replace .color() with .foregroundColor() to reflect the changes in beta 4
+
 ## [0.2.0] - 2019-07-12
 - [Added] Static components
 - [Changed] Rename AttrText's typealias to AText